+++ app/services/admin_conversation_state.go
@@ -6,12 +6,12 @@ import (
 	"net/http"
 )
 
-func (s *appService) GetAdminConversationState(adminID uint8, conversationID uint64) (*models.AdminConversationState, error) {
+func (s *appService) GetAdminConversationState(adminID uint64, conversationID uint64) (*models.AdminConversationState, error) {
 	return s.repo.GetAdminConversationState(adminID, conversationID)
 }
 
 func (s *appService) GetAdminListConversationStates(claim models.User) helpers.Response {
-	states, err := s.repo.GetAdminConversationStatesByAdminID(uint8(claim.ID))
+	states, err := s.repo.GetAdminConversationStatesByAdminID(claim.ID)
 	if err != nil {
 		return helpers.NewResponse(http.StatusInternalServerError, "Failed to retrieve conversation states", nil, nil)
 	}
